lvol: switch from is_healthy to is_degraded
Subsequent to the backport of external snapshots from v23.01 development
to v23.01.nvda, the way that missing external snapshots were reported
changed. Previously an lvol that was broken due to a missing external
snapshot was reported in bdev_get_bdevs with is_healthy set to false.
Now it is reported as is_degraded set to true.

Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>