test/lvol/esnap: resync with upstream
This updates tests to match spdk upstream. Without these updates the
tests are usually successful on a fast bare metal machine with fast
storage but usually fail when run in a VM, presumably due to slightly
higher aio latency. This is fixed with wrapping poll_threads() in
poll_error_updated().

Signed-off-by: Mike Gerdts <mgerdts@nvidia.com>